+++ NEWS.md
@@ -2,6 +2,8 @@
 
 ## New features
 
+* Added a plot function -- `plot.ranef.mjoint()` -- for `ranef.mjoint` objects.
+
 * `bootSE()` now uses control parameters from fitted model and allows for individual parameter overwrite.
 
 * Added a check that any initial covariance matrix given is positive-definite.